% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODIFICADOSCM.R
\name{CODIFICADOSCM}
\alias{CODIFICADOSCM}
\title{CODIFICAR SOLAPAS DE  CADA TIPO DE PROYECTO}
\usage{
CODIFICADOSCM(MUNI, NPRO)
}
\arguments{
\item{1}{Los parametros a utilizar son los siguientes: MUNI = MUNICIPIO (en mayuscula),
NPRO = numero de proyecto}
}
\description{
Esta funcion esta creada para codificar la solapa de SC de una UG municipal
Requiere algunos pre requisitos: La columna de N de proyecto tiene que estar descombinada y todas
las lineas deben tener el n de proyecto para filtrar correctamente.
También debemos tener cargados algunos objetos previo a la utilización de esta funcio:
}
\details{
INDEC: la solapa con los códigos del INDEC
EXP: Listado de las UG con sus nombre y nro de expedientes
TABlA : La tabla con los códigos de tipo de proyectos y Ejes
CODSC : Solapa de proyectos socio comunitarios
}
\examples{
CODIFICAR(MORENO, i) 





}
\keyword{Codificacion,}
\keyword{Potenciar}
\keyword{trabajo}
