% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODIFICADOSPM.R
\name{CODIFICADOSPM}
\alias{CODIFICADOSPM}
\title{CODIFICAR SOLAPAS DE  CADA TIPO DE PROYECTO}
\usage{
CODIFICADOSPM(MUNI, NPRO)
}
\arguments{
\item{1}{Los parametros a utilizar son los siguientes: MUNI = MUNICIPIO (en mayuscula),
NPRO = numero de proyecto}
}
\description{
Esta funcion esta creada para codificar la solapa de SP de una UG municipal
Requiere algunos pre requisitos: La columna de N de proyecto tiene que estar descombinada y todas
las lineas deben tener el n de proyecto para filtrar correctamente.
Tambien debemos tener cargados algunos objetos previo a la utilizacion de esta funcio:
}
\details{
INDEC: la solapa con los codigos del INDEC
EXP: Listado de las UG con sus nombre y nro de expedientes
TABlA : La tabla con los códigos de tipo de proyectos y Ejes
CODSP : Solapa de proyectos socio  productivos
}
\examples{
CODIFICAR(MORENO, i) 




}
\keyword{Codificacion,}
\keyword{Potenciar}
\keyword{trabajo}
